utils/frame_queue: add deinterlacing support
After a lot of internal debating and back-and-forth, I decided to settle
on an approach where the pl_queue splits up frames internally. This
frees the user from the burden of having to deal with these relatively
common corner cases, and also allows us to try and be as smart as
possible and 'do the right thing' with respect to things like doubling
detected framerate, pushing frames out-of-order, re-linking past and
future frames in all sorts of corner cases, and internally refcounting
frames on an actual as-needed basis.

This approach is relatively transparent for users and merely requires
them forward the appropriate field metadata.

It's worth pointing out that the current design of `struct entry` sort
of conflates two separate concepts: source frames, and their
corresponding fields/queue entries pointing at them. The way this is
organized internally is currently sort of ad-hoc and only done this way
to save a bit of memory and convenience. I may decide to reorganize it
in the future.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/218